Fix Json serialization.
