+++ src/client/Microsoft.Identity.Client.Services/DeviceCodeResult.cs
