+++ src/client/Microsoft.Identity.Client.Services/Internal/DeviceCodeResponse.cs
