+++ src/client/Microsoft.Identity.Client.Services/OAuth2/MsalTokenResponse.cs
@@ -8,7 +8,6 @@ using System.Text;
 using Microsoft.Identity.Client.Core;
 using Microsoft.Identity.Client.Extensibility;
 using Microsoft.Identity.Client.Http;
-using Microsoft.Identity.Client.Internal.Broker;
 using Microsoft.Identity.Client.Utils;
 #if SUPPORTS_SYSTEM_TEXT_JSON
 using System.Text.Json;
@@ -55,8 +54,6 @@ namespace Microsoft.Identity.Client.OAuth2
 
         }
 
-        private const string iOSBrokerErrorMetadata = "error_metadata";
-        private const string iOSBrokerHomeAccountId = "home_account_id";
         [JsonProperty(TokenResponseClaim.TokenType)]
         public string TokenType { get; set; }
 
@@ -108,70 +105,6 @@ namespace Microsoft.Identity.Client.OAuth2
 
         public HttpResponse HttpResponse { get; set; }
 
-        internal static MsalTokenResponse CreateFromiOSBrokerResponse(Dictionary<string, string> responseDictionary)
-        {
-            if (responseDictionary.TryGetValue(BrokerResponseConst.BrokerErrorCode, out string errorCode))
-            {
-                string metadataOriginal = responseDictionary.ContainsKey(MsalTokenResponse.iOSBrokerErrorMetadata) ? responseDictionary[MsalTokenResponse.iOSBrokerErrorMetadata] : null;
-                Dictionary<string, string> metadataDictionary = null;
-
-                if (metadataOriginal != null)
-                {
-                    string brokerMetadataJson = Uri.UnescapeDataString(metadataOriginal);
-#if SUPPORTS_SYSTEM_TEXT_JSON
-                    metadataDictionary = new Dictionary<string, string>();
-                    foreach (var item in JsonDocument.Parse(brokerMetadataJson).RootElement.EnumerateObject())
-                    {
-                        metadataDictionary.Add(item.Name, item.Value.GetString());
-                    }
-#else
-                    metadataDictionary = JsonConvert.DeserializeObject<Dictionary<string, string>>(brokerMetadataJson);
-#endif
-                }
-
-                string homeAcctId = null;
-                metadataDictionary?.TryGetValue(MsalTokenResponse.iOSBrokerHomeAccountId, out homeAcctId);
-                return new MsalTokenResponse
-                {
-                    Error = errorCode,
-                    ErrorDescription = responseDictionary.ContainsKey(BrokerResponseConst.BrokerErrorDescription) ? CoreHelpers.UrlDecode(responseDictionary[BrokerResponseConst.BrokerErrorDescription]) : string.Empty,
-                    SubError = responseDictionary.ContainsKey(OAuth2ResponseBaseClaim.SubError) ? responseDictionary[OAuth2ResponseBaseClaim.SubError] : string.Empty,
-                    AccountUserId = homeAcctId != null ? AccountId.ParseFromString(homeAcctId).ObjectId : null,
-                    TenantId = homeAcctId != null ? AccountId.ParseFromString(homeAcctId).TenantId : null,
-                    Upn = (metadataDictionary?.ContainsKey(TokenResponseClaim.Upn) ?? false) ? metadataDictionary[TokenResponseClaim.Upn] : null,
-                    CorrelationId = responseDictionary.ContainsKey(BrokerResponseConst.CorrelationId) ? responseDictionary[BrokerResponseConst.CorrelationId] : null,
-                };
-            }
-
-            var response = new MsalTokenResponse
-            {
-                AccessToken = responseDictionary[BrokerResponseConst.AccessToken],
-                RefreshToken = responseDictionary.ContainsKey(BrokerResponseConst.RefreshToken)
-                    ? responseDictionary[BrokerResponseConst.RefreshToken]
-                    : null,
-                IdToken = responseDictionary[BrokerResponseConst.IdToken],
-                TokenType = BrokerResponseConst.Bearer,
-                CorrelationId = responseDictionary[BrokerResponseConst.CorrelationId],
-                Scope = responseDictionary[BrokerResponseConst.Scope],
-                ExpiresIn = responseDictionary.TryGetValue(BrokerResponseConst.ExpiresOn, out string expiresOn) ?
-                                DateTimeHelpers.GetDurationFromNowInSeconds(expiresOn) :
-                                0,
-                ClientInfo = responseDictionary.ContainsKey(BrokerResponseConst.ClientInfo)
-                                ? responseDictionary[BrokerResponseConst.ClientInfo]
-                                : null,
-                TokenSource = TokenSource.Broker
-            };
-
-            if (responseDictionary.ContainsKey(TokenResponseClaim.RefreshIn))
-            {
-                response.RefreshIn = long.Parse(
-                    responseDictionary[TokenResponseClaim.RefreshIn],
-                    CultureInfo.InvariantCulture);
-            }
-
-            return response;
-        }
-
         internal static MsalTokenResponse CreateFromAppProviderResponse(AppTokenProviderResult tokenProviderResponse)
         {
             ValidateTokenProviderResult(tokenProviderResponse);
@@ -181,7 +114,7 @@ namespace Microsoft.Identity.Client.OAuth2
                 AccessToken = tokenProviderResponse.AccessToken,
                 RefreshToken = null,
                 IdToken = null,
-                TokenType = BrokerResponseConst.Bearer,
+                TokenType = "Bearer",
                 ExpiresIn = tokenProviderResponse.ExpiresInSeconds,
                 ClientInfo = null,
                 TokenSource = TokenSource.IdentityProvider,
@@ -209,50 +142,6 @@ namespace Microsoft.Identity.Client.OAuth2
         private static void HandleInvalidExternalValueError(string nameOfValue)
         {
             throw new MsalClientException(MsalError.InvalidTokenProviderResponseValue, MsalErrorMessage.InvalidTokenProviderResponseValue(nameOfValue));
-        }
-
-        /// <remarks>
-        /// This method does not belong here - it is more tied to the Android code. However, that code is
-        /// not unit testable, and this one is. 
-        /// The values of the JSON response are based on 
-        /// https://github.com/AzureAD/microsoft-authentication-library-common-for-android/blob/dev/common/src/main/java/com/microsoft/identity/common/internal/broker/BrokerResult.java
-        /// </remarks>
-        internal static MsalTokenResponse CreateFromAndroidBrokerResponse(string jsonResponse, string correlationId)
-        {
-            var authResult = JsonHelper.ParseIntoJsonObject(jsonResponse);
-            var errorCode = authResult[BrokerResponseConst.BrokerErrorCode]?.ToString();
-
-            if (!string.IsNullOrEmpty(errorCode))
-            {
-                return new MsalTokenResponse
-                {
-                    Error = errorCode,
-                    ErrorDescription = authResult[BrokerResponseConst.BrokerErrorMessage]?.ToString(),
-                    AuthorityUrl = authResult[BrokerResponseConst.Authority]?.ToString(),
-                    TenantId = authResult[BrokerResponseConst.TenantId]?.ToString(),
-                    Upn = authResult[BrokerResponseConst.UserName]?.ToString(),
-                    AccountUserId = authResult[BrokerResponseConst.LocalAccountId]?.ToString(),
-                };
-            }
-
-            MsalTokenResponse msalTokenResponse = new MsalTokenResponse()
-            {
-                AccessToken = authResult[BrokerResponseConst.AccessToken].ToString(),
-                IdToken = authResult[BrokerResponseConst.IdToken].ToString(),
-                CorrelationId = correlationId, // Android response does not expose Correlation ID
-                Scope = authResult[BrokerResponseConst.AndroidScopes].ToString(), // sadly for iOS this is "scope" and for Android "scopes"
-                ExpiresIn = DateTimeHelpers.GetDurationFromNowInSeconds(authResult[BrokerResponseConst.ExpiresOn].ToString()),
-                ExtendedExpiresIn = DateTimeHelpers.GetDurationFromNowInSeconds(authResult[BrokerResponseConst.ExtendedExpiresOn].ToString()),
-                ClientInfo = authResult[BrokerResponseConst.ClientInfo].ToString(),
-                TokenType = authResult[BrokerResponseConst.TokenType]?.ToString() ?? "Bearer",
-                TokenSource = TokenSource.Broker,
-                AuthorityUrl = authResult[BrokerResponseConst.Authority]?.ToString(),
-                TenantId = authResult[BrokerResponseConst.TenantId]?.ToString(),
-                Upn = authResult[BrokerResponseConst.UserName]?.ToString(),
-                AccountUserId = authResult[BrokerResponseConst.LocalAccountId]?.ToString(),
-            };
-
-            return msalTokenResponse;
         }
 
         public void Log(ILoggerAdapter logger, LogLevel logLevel)
