+++ src/client/Microsoft.Identity.Client.Services/WsTrust/UserRealmDiscoveryResponse.cs
